services/asr/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
nats-py==2.6.0
websockets==12.0
minio==7.2.7
pydantic==2.9.2
# torch>=2.0.0  # Uncomment when integrating models
# torchaudio>=2.0.0
# openai-whisper>=20231117  # Uncomment when using Whisper
